+++ Task_02/Task_02.csproj
@@ -47,7 +47,7 @@
     <Compile Include="Entities\Hero.cs" />
     <Compile Include="Enums\HeroNames.cs" />
     <Compile Include="Entities\HeroScoreInfo.cs" />
-    <Compile Include="HeroScoreInfosFactory.cs" />
+    <Compile Include="Factories\HeroScoreInfosFactory.cs" />
     <Compile Include="Statistics\Abstract\NumberOfMatches\HeroUsagesStatisticMostFavoriteProvider.cs" />
     <Compile Include="Statistics\Abstract\WinRate\HeroUsagesStatisticMostSuccesfulProvider.cs" />
     <Compile Include="Printers\Certain\HeroUsagesStatisticPrinter.cs" />
@@ -57,7 +57,7 @@
     <Compile Include="Statistics\Certain\HeroUsagesStatisticProvider.cs" />
     <Compile Include="Statistics\Abstract\HeroUsagesStatisticWinStreakProvider.cs" />
     <Compile Include="Printers\Abstract\IPrinter.cs" />
-    <Compile Include="ScoreInfoCalculator.cs" />
+    <Compile Include="Entities\ScoreInfoCalculator.cs" />
     <Compile Include="Printers\Abstract\NumberOfMatches\FavoritePrinter.cs" />
     <Compile Include="Printers\Abstract\NumberOfMatches\NumberOfMatchesPrinter.cs" />
     <Compile Include="Printers\Abstract\NumberOfMatches\UnfavoritePrinter.cs" />
